trying to get it work

+++ src/training.py
@@ -225,8 +225,8 @@ def loss_recon(recon_x, x):
     return recon_loss
 
 def distance_from_mean(recon_x, mean_distribution):
-        #recon_x = recon_x.double()
-        #mean_distribution = mean_distribution.double()
+        recon_x = recon_x.double()
+        mean_distribution = mean_distribution.double()
         mean_distance = 1 -  torch.abs(CCC_loss(recon_x, mean_distribution))
         return mean_distance
